requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyMySQL==1.1.1
pyngrok==7.2.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
